requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-filter==2.2.0
django-jsonfield==1.3.1
djangorestframework==3.11.0
idna==2.8
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
zipfile36==0.1.3
